* Improve centering * Browser-less support * Move to utility function
